Add language annotations, fix Makefile, CI
