+++ docker-compose.medea.yml
@@ -5,8 +5,7 @@ services:
     container_name: ${COMPOSE_PROJECT_NAME}-backend
     image: ${COMPOSE_IMAGE_NAME}:${COMPOSE_IMAGE_VER}
     build: .
-    depends_on:
-      - coturn
+    depends_on: ["coturn"]
     environment:
       RUST_LOG: ${RUST_LOG}
       MEDEA_CONF: ${MEDEA_CONF}
