+++ tests/e2e/nginx.conf
@@ -16,7 +16,7 @@ http {
             root   /usr/share/nginx/html;
             index  index.html index.htm;
             types {
-                application/javascript js mjs;
+                application/javascript js;
                 application/wasm wasm;
                 text/html html;
             }
