+++ tests/e2e/object/jason.rs
@@ -12,6 +12,7 @@ pub struct Jason;
 impl Builder for Jason {
     fn build(self) -> Statement {
         Statement::new(
+            // language=JavaScript
             r#"
                 async () => {
                     let jason = new window.rust.Jason();
@@ -27,6 +28,7 @@ impl Object<Jason> {
     /// Returns a new [`Room`] initiated in this [`Jason`].
     pub async fn init_room(&self) -> Result<Object<Room>, super::Error> {
         self.execute_and_fetch(Statement::new(
+            // language=JavaScript
             r#"
                 async (jason) => {
                     let room = await jason.init_room();
